o stores now closed properly when converting from old store to property index store o changed some checks in PropertyRecord to assertions o fixed a int multiplier overflow
git-svn-id: https://svn.neo4j.org/projects/neo/trunk@393 0b971d98-bb2f-0410-8247-b05b2b5feb2a

+++ advanced/src/java/org/neo4j/impl/nioneo/store/AbstractDynamicStore.java
@@ -503,7 +503,7 @@ public abstract class AbstractDynamicStore extends CommonAbstractStore
 			if ( buffer != null && !hasWindow( blockId ) )
 			{
 				buffer.makeReadyForTransfer();
-				getFileChannel().transferTo( blockId * getBlockSize(), 
+				getFileChannel().transferTo( (long) blockId * getBlockSize(), 
 					getBlockSize(), buffer.getFileChannel() );
 				ByteBuffer buf = buffer.getByteBuffer();
 				record = new DynamicRecord( blockId );
